HwmonTempMain: revert the type of bus and address back to uint64_t
In I546e967abae7c0fb9fca645867e3037046037647, the type of bus and
address in HwmonTempMain were updated to size_t.

However, in some systems using i3c, the address can be greater than the
range of size_t. Thus, reverting the type back to uint64_t

Tested:
- added unit test passed.
- can read temperature of the i3c devices from our system.

+++ src/HwmonTempMain.cpp
@@ -302,8 +302,8 @@ void createSensors(
                 deviceName = fs::canonical(device).stem();
             }
 
-            size_t bus = 0;
-            size_t addr = 0;
+            uint64_t bus = 0;
+            uint64_t addr = 0;
             if (!getDeviceBusAddr(deviceName, bus, addr))
             {
                 continue;
